requirements.txt
# MediGuide Backend Dependencies

# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.9

# Supabase
supabase==2.9.0
# postgrest handled by supabase


# Authentication & Security
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
pydantic-settings==2.7.1
python-dotenv==1.0.0

# AI/ML
openai==1.12.0

# OCR
pytesseract==0.3.10
Pillow==11.0.0

# Image processing
opencv-python-headless==4.9.0.80

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pydantic==2.10.4

# Optional: For production background tasks
# celery==5.3.4
# redis==5.0.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3

